+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/mapping/ThreadMapping.kt
@@ -28,7 +28,7 @@ internal fun DownstreamThreadDto.toDomain(currentUserId: UserId?): Thread =
     Thread(
         activeParticipantCount = active_participant_count ?: 0,
         cid = channel_cid,
-        channel = channel?.toDomain(currentUserId),
+        channel = channel?.toDomain(currentUserId, null),
         parentMessageId = parent_message_id,
         parentMessage = parent_message.toDomain(currentUserId),
         createdByUserId = created_by_user_id,