+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/parser/EventArguments.kt
@@ -151,6 +151,7 @@ internal object EventArguments {
 
     private val member = Member(
         user,
+        channelRole = "channel_member",
         createdAt = date,
         updatedAt = date,
     )
@@ -188,7 +189,7 @@ internal object EventArguments {
     private val channel = Channel(
         id = channelId,
         type = channelType,
-        lastMessageAt = date,
+        channelLastMessageAt = date,
         createdAt = date,
         updatedAt = date,
         createdBy = user,
@@ -226,6 +227,7 @@ internal object EventArguments {
         channelType = channelType,
         channelId = channelId,
         channel = channel,
+        channelLastMessageAt = date,
         user = user,
     )
 
@@ -237,6 +239,7 @@ internal object EventArguments {
         channelType = channelType,
         channelId = channelId,
         user = user,
+        channelLastMessageAt = date,
         clearHistory = true,
     )
 
@@ -247,9 +250,10 @@ internal object EventArguments {
         cid = cid,
         channelType = channelType,
         channelId = channelId,
+        channel = channel,
+        channelLastMessageAt = date,
         user = user,
         message = null,
-        channel = channel,
     )
     private val channelTruncatedServerSideEvent = ChannelTruncatedEvent(
         type = EventType.CHANNEL_TRUNCATED,
@@ -258,9 +262,10 @@ internal object EventArguments {
         cid = cid,
         channelType = channelType,
         channelId = channelId,
+        channel = channel,
+        channelLastMessageAt = date,
         user = null,
         message = null,
-        channel = channel,
     )
     private val channelUpdatedEvent = ChannelUpdatedEvent(
         type = EventType.CHANNEL_UPDATED,
@@ -271,6 +276,7 @@ internal object EventArguments {
         channelId = channelId,
         message = message,
         channel = channel,
+        channelLastMessageAt = date,
     )
     private val channelUpdatedByUserEvent = ChannelUpdatedByUserEvent(
         type = EventType.CHANNEL_UPDATED,
@@ -280,8 +286,9 @@ internal object EventArguments {
         channelType = channelType,
         channelId = channelId,
         user = user,
-        message = message,
         channel = channel,
+        channelLastMessageAt = date,
+        message = message,
     )
     private val channelVisibleEvent = ChannelVisibleEvent(
         type = EventType.CHANNEL_VISIBLE,
@@ -291,6 +298,7 @@ internal object EventArguments {
         channelType = channelType,
         channelId = channelId,
         user = user,
+        channelLastMessageAt = date,
     )
     private val memberAddedEvent = MemberAddedEvent(
         type = EventType.MEMBER_ADDED,
@@ -301,6 +309,7 @@ internal object EventArguments {
         channelType = channelType,
         channelId = channelId,
         member = member,
+        channelLastMessageAt = date,
     )
     private val memberRemovedEvent = MemberRemovedEvent(
         type = EventType.MEMBER_REMOVED,
@@ -311,6 +320,7 @@ internal object EventArguments {
         channelType = channelType,
         channelId = channelId,
         member = member,
+        channelLastMessageAt = date,
     )
     private val memberUpdatedEvent = MemberUpdatedEvent(
         type = EventType.MEMBER_UPDATED,
@@ -321,6 +331,7 @@ internal object EventArguments {
         channelType = channelType,
         channelId = channelId,
         member = member,
+        channelLastMessageAt = date,
     )
     private val messageDeletedEvent = MessageDeletedEvent(
         type = EventType.MESSAGE_DELETED,
@@ -330,7 +341,9 @@ internal object EventArguments {
         cid = cid,
         channelType = channelType,
         channelId = channelId,
-        message = message, hardDelete = false,
+        message = message,
+        hardDelete = false,
+        channelLastMessageAt = date,
     )
     private val messageDeletedServerSideEvent = MessageDeletedEvent(
         type = EventType.MESSAGE_DELETED,
@@ -340,7 +353,9 @@ internal object EventArguments {
         cid = cid,
         channelType = channelType,
         channelId = channelId,
-        message = message, hardDelete = true,
+        message = message,
+        hardDelete = true,
+        channelLastMessageAt = date,
     )
     private val messageReadEvent = MessageReadEvent(
         type = EventType.MESSAGE_READ,
@@ -350,6 +365,7 @@ internal object EventArguments {
         cid = cid,
         channelType = channelType,
         channelId = channelId,
+        channelLastMessageAt = date,
     )
     private val messageUpdatedEvent = MessageUpdatedEvent(
         type = EventType.MESSAGE_UPDATED,
@@ -360,6 +376,7 @@ internal object EventArguments {
         channelType = channelType,
         channelId = channelId,
         message = message,
+        channelLastMessageAt = date,
     )
     private val notificationAddedToChannelEvent = NotificationAddedToChannelEvent(
         type = EventType.NOTIFICATION_ADDED_TO_CHANNEL,
@@ -372,6 +389,7 @@ internal object EventArguments {
         member = member,
         totalUnreadCount = totalUnreadCount,
         unreadChannels = unreadChannels,
+        channelLastMessageAt = date,
     )
     private val notificationChannelDeletedEvent = NotificationChannelDeletedEvent(
         type = EventType.NOTIFICATION_CHANNEL_DELETED,
@@ -381,6 +399,7 @@ internal object EventArguments {
         channelType = channelType,
         channelId = channelId,
         channel = channel,
+        channelLastMessageAt = date,
     )
     private val notificationChannelTruncatedEvent = NotificationChannelTruncatedEvent(
         type = EventType.NOTIFICATION_CHANNEL_TRUNCATED,
@@ -390,6 +409,7 @@ internal object EventArguments {
         channelType = channelType,
         channelId = channelId,
         channel = channel,
+        channelLastMessageAt = date,
     )
     private val notificationInviteAcceptedEvent = NotificationInviteAcceptedEvent(
         type = EventType.NOTIFICATION_INVITE_ACCEPTED,
@@ -401,6 +421,7 @@ internal object EventArguments {
         user = user,
         member = member,
         channel = channel,
+        channelLastMessageAt = date,
     )
     private val notificationInviteRejectedEvent = NotificationInviteRejectedEvent(
         type = EventType.NOTIFICATION_INVITE_REJECTED,
@@ -412,6 +433,7 @@ internal object EventArguments {
         user = user,
         member = member,
         channel = channel,
+        channelLastMessageAt = date,
     )
     private val notificationInvitedEvent = NotificationInvitedEvent(
         type = EventType.NOTIFICATION_INVITED,
@@ -422,6 +444,7 @@ internal object EventArguments {
         channelId = channelId,
         user = user,
         member = member,
+        channelLastMessageAt = date,
     )
     private val notificationMarkReadEvent = NotificationMarkReadEvent(
         type = EventType.NOTIFICATION_MARK_READ,
@@ -433,6 +456,7 @@ internal object EventArguments {
         channelId = channelId,
         totalUnreadCount = totalUnreadCount,
         unreadChannels = unreadChannels,
+        channelLastMessageAt = date,
     )
     private val notificationMarkUnreadEvent = NotificationMarkUnreadEvent(
         type = EventType.NOTIFICATION_MARK_UNREAD,
@@ -448,6 +472,7 @@ internal object EventArguments {
         firstUnreadMessageId = message.id,
         lastReadMessageAt = date,
         lastReadMessageId = parentMessageId,
+        channelLastMessageAt = date,
     )
     private val notificationMessageNewEvent = NotificationMessageNewEvent(
         type = EventType.NOTIFICATION_MESSAGE_NEW,
@@ -460,6 +485,7 @@ internal object EventArguments {
         message = message,
         totalUnreadCount = totalUnreadCount,
         unreadChannels = unreadChannels,
+        channelLastMessageAt = date,
     )
     private val notificationRemovedFromChannelEvent = NotificationRemovedFromChannelEvent(
         type = EventType.NOTIFICATION_REMOVED_FROM_CHANNEL,
@@ -471,6 +497,7 @@ internal object EventArguments {
         channelId = channelId,
         channel = channel,
         member = member,
+        channelLastMessageAt = date,
     )
     private val reactionDeletedEvent = ReactionDeletedEvent(
         type = EventType.REACTION_DELETED,
@@ -482,6 +509,7 @@ internal object EventArguments {
         channelId = channelId,
         message = message,
         reaction = reaction,
+        channelLastMessageAt = date,
     )
     private val reactionNewEvent = ReactionNewEvent(
         type = EventType.REACTION_NEW,
@@ -493,6 +521,7 @@ internal object EventArguments {
         channelId = channelId,
         message = message,
         reaction = reaction,
+        channelLastMessageAt = date,
     )
     private val reactionUpdateEvent = ReactionUpdateEvent(
         type = EventType.REACTION_UPDATED,
@@ -504,6 +533,7 @@ internal object EventArguments {
         channelId = channelId,
         message = message,
         reaction = reaction,
+        channelLastMessageAt = date,
     )
     private val typingStartEvent = TypingStartEvent(
         type = EventType.TYPING_START,
@@ -514,6 +544,7 @@ internal object EventArguments {
         channelType = channelType,
         channelId = channelId,
         parentId = parentMessageId,
+        channelLastMessageAt = date,
     )
     private val typingStopEvent = TypingStopEvent(
         type = EventType.TYPING_STOP,
@@ -524,6 +555,7 @@ internal object EventArguments {
         channelType = channelType,
         channelId = channelId,
         parentId = parentMessageId,
+        channelLastMessageAt = date,
     )
     private val channelUserBannedEvent = ChannelUserBannedEvent(
         type = EventType.USER_BANNED,
@@ -535,6 +567,7 @@ internal object EventArguments {
         user = user,
         expiration = date,
         shadow = false,
+        channelLastMessageAt = date,
     )
     private val globalUserBannedEvent = GlobalUserBannedEvent(
         type = EventType.USER_BANNED,
@@ -563,6 +596,7 @@ internal object EventArguments {
         channelType = channelType,
         channelId = channelId,
         user = user,
+        channelLastMessageAt = date,
     )
     private val userStopWatchingEvent = UserStopWatchingEvent(
         type = EventType.USER_WATCHING_STOP,
@@ -573,6 +607,7 @@ internal object EventArguments {
         channelType = channelType,
         channelId = channelId,
         user = user,
+        channelLastMessageAt = date,
     )
     private val channelUserUnbannedEvent = ChannelUserUnbannedEvent(
         type = EventType.USER_UNBANNED,
@@ -582,6 +617,7 @@ internal object EventArguments {
         cid = cid,
         channelType = channelType,
         channelId = channelId,
+        channelLastMessageAt = date,
     )
     private val globalUserUnbannedEvent = GlobalUserUnbannedEvent(
         type = EventType.USER_UNBANNED,
@@ -632,6 +668,7 @@ internal object EventArguments {
         watcherCount = watcherCount,
         totalUnreadCount = totalUnreadCount,
         unreadChannels = unreadChannels,
+        channelLastMessageAt = date,
     )
     private val newMessageWithoutUnreadCountsEvent = NewMessageEvent(
         type = EventType.MESSAGE_NEW,
@@ -643,6 +680,7 @@ internal object EventArguments {
         channelId = channelId,
         message = message,
         watcherCount = watcherCount,
+        channelLastMessageAt = date,
     )
     private val unknownEvent = UnknownEvent(
         type = EventType.UNKNOWN,