+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/parser2/testdata/ChannelDtoTestData.kt
@@ -47,6 +47,7 @@ internal object ChannelDtoTestData {
           "url_enrichment": false,
           "custom_events": false,
           "push_notifications": true,
+          "skip_last_msg_update_for_system_msgs": false,
           "polls": true,
           "message_retention": "retention",
           "max_message_length": 500,
@@ -79,6 +80,7 @@ internal object ChannelDtoTestData {
         url_enrichment = false,
         custom_events = false,
         push_notifications = true,
+        skip_last_msg_update_for_system_msgs = false,
         polls = true,
         message_retention = "retention",
         max_message_length = 500,
@@ -114,7 +116,6 @@ internal object ChannelDtoTestData {
           "members": [
            {
             "user": ${UserDtoTestData.downstreamJson},
-            "role": "owner",
             "created_at": "2020-06-10T11:04:31.0Z",
             "updated_at": "2020-06-10T11:04:31.588Z",
             "invited": true,
@@ -145,7 +146,6 @@ internal object ChannelDtoTestData {
           "own_capabilities": ["connect-events", "pin-message"],
           "membership": {
             "user": ${UserDtoTestData.downstreamJson},
-            "role": "owner",
             "created_at": "2020-06-10T11:04:31.0Z",
             "updated_at": "2020-06-10T11:04:31.588Z",
             "invited": true,
@@ -186,6 +186,7 @@ internal object ChannelDtoTestData {
                 notifications_muted = false,
                 status = "member",
                 ban_expires = Date(1615218151355),
+                extraData = emptyMap(),
             ),
         ),
         watchers = listOf(UserDtoTestData.downstreamUser),
@@ -216,6 +217,7 @@ internal object ChannelDtoTestData {
             notifications_muted = false,
             status = "member",
             ban_expires = null,
+            extraData = emptyMap(),
         ),
         extraData = mapOf("draft" to true),
     )
@@ -371,6 +373,7 @@ internal object ChannelDtoTestData {
                 notifications_muted = false,
                 status = "member",
                 ban_expires = Date(1615218151355),
+                extraData = emptyMap(),
             ),
         ),
         watchers = listOf(UserDtoTestData.upstreamUser),