+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add `ChatClient::partialUpdateMember` for updating a `Member` in the scope of a `Channel`. [#5497](https://github.com/GetStream/stream-chat-android/pull/5497)
+- Add `ChannelClient::partialUpdateMember` for updating a `Member` in the scope of a `Channel`. [#5497](https://github.com/GetStream/stream-chat-android/pull/5497)
 
 ### ⚠️ Changed
 
@@ -47,6 +49,8 @@
 ## stream-chat-android-ui-common
 ### 🐞 Fixed
 - Fix wrong timestamp used for the initial loading of pinned messages in `PinnedMessageListController`. [#5483](https://github.com/GetStream/stream-chat-android/pull/5483)
+- Add `DownloadAttachmentUriGenerator` to generate download URIs for attachments. [#5490](https://github.com/GetStream/stream-chat-android/pull/5490)
+- Add `DownloadRequestInterceptor` to intercept download requests. [#5490](https://github.com/GetStream/stream-chat-android/pull/5490)
 
 ### ⬆️ Improved
 - Improve performance of opening a thread or message in a thread via `MessageListController`. [#5489](https://github.com/GetStream/stream-chat-android/pull/5489)
@@ -63,6 +67,8 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add `ChatUI.downloadAttachmentUriGenerator` property to generate download URIs for attachments. [#5490](https://github.com/GetStream/stream-chat-android/pull/5490)
+- Add `ChatUI.downloadRequestInterceptor` property to intercept download requests. [#5490](https://github.com/GetStream/stream-chat-android/pull/5490)
 
 ### ⚠️ Changed
 
@@ -74,6 +80,8 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Add `ChatTheme.downloadAttachmentUriGenerator` property to generate download URIs for attachments. [#5490](https://github.com/GetStream/stream-chat-android/pull/5490)
+- Add `ChatTheme.downloadRequestInterceptor` property to intercept download requests. [#5490](https://github.com/GetStream/stream-chat-android/pull/5490)
 
 ### ⚠️ Changed